+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/exac/ExacAnnotationDriverReportOverlappingAsMatchingTest.java
@@ -96,8 +96,8 @@ public class ExacAnnotationDriverReportOverlappingAsMatchingTest extends ExacAnn
 		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_HEMI_ALL"));
 		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_HET_ALL"));
 
-		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_BEST_AC"));
-		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_BEST_AF"));
+		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_AC_POPMAX"));
+		Assert.assertNotNull(header.getInfoHeaderLine("EXAC_AF_POPMAX"));
 	}
 
 	@Test
@@ -115,9 +115,9 @@ public class ExacAnnotationDriverReportOverlappingAsMatchingTest extends ExacAnn
 		Assert.assertEquals(42, annotated.getAttributes().size());
 		ArrayList<String> keys = Lists.newArrayList(annotated.getAttributes().keySet());
 		Collections.sort(keys);
-		Assert.assertEquals("[AC_AFR, AC_ALL, AC_AMR, AC_EAS, AC_FIN, AC_NFE, AC_OTH, AC_SAS, AF_AFR, AF_ALL, "
-				+ "AF_AMR, AF_EAS, AF_FIN, AF_NFE, AF_OTH, AF_SAS, AN_AFR, AN_ALL, AN_AMR, "
-				+ "AN_EAS, AN_FIN, AN_NFE, AN_OTH, AN_SAS, BEST_AC, BEST_AF, "
+		Assert.assertEquals("[AC_AFR, AC_ALL, AC_AMR, AC_EAS, AC_FIN, AC_NFE, AC_OTH, AC_POPMAX, AC_SAS, AF_AFR, AF_ALL, "
+				+ "AF_AMR, AF_EAS, AF_FIN, AF_NFE, AF_OTH, AF_POPMAX, AF_SAS, AN_AFR, AN_ALL, AN_AMR, "
+				+ "AN_EAS, AN_FIN, AN_NFE, AN_OTH, AN_SAS, "
 				+ "HET_AFR, HET_ALL, HET_AMR, HET_EAS, HET_FIN, HET_NFE, HET_OTH, HET_SAS, "
 				+ "HOM_AFR, HOM_ALL, HOM_AMR, HOM_EAS, HOM_FIN, HOM_NFE, HOM_OTH, HOM_SAS]", keys.toString());
 
@@ -150,9 +150,9 @@ public class ExacAnnotationDriverReportOverlappingAsMatchingTest extends ExacAnn
 		Assert.assertEquals("122", annotated.getAttributeAsString("AN_OTH", null));
 		Assert.assertEquals("7520", annotated.getAttributeAsString("AN_SAS", null));
 
-		Assert.assertEquals("[2, 2, 2]", annotated.getAttributeAsString("BEST_AC", null));
+		Assert.assertEquals("[2, 2, 2]", annotated.getAttributeAsString("AC_POPMAX", null));
 		Assert.assertEquals("[0.003194888178913738, 0.003194888178913738, 0.003194888178913738]",
-				annotated.getAttributeAsString("BEST_AF", null));
+				annotated.getAttributeAsString("AF_POPMAX", null));
 	}
 
 }